feat: 카트->주문시 라이브쇼핑/상품홍보 정보 올바르게 입력되도록 구성 + 코드정리
